Refactor db sanity checks to accommodate for multiple sanity check functions:
* Rename 'slotRelatedChecks' -> 'slotAttendance'
* Extract out 'api' functions (as they're to be reused for different sanity check functionalities)
