+++ packages/functions/src/checks/slotAttendance.ts
@@ -1,21 +1,19 @@
+import { DateTime } from "luxon";
 import admin from "firebase-admin";
 
 import {
   AttendanceAutofixReport,
   Collection,
-  DateMismatchDoc,
-  FirestoreSchema,
   OrgSubCollection,
-  SanityCheckKind,
   SlotAttendanceUpdate,
   SlotAttendnace,
   SlotInterface,
-  SlotSanityCheckReport,
-  UnpairedDoc,
+  SlotAttendanceSanityCheckReport,
   map,
+  DateMismatchDoc,
+  UnpairedDoc,
   wrapIter,
 } from "@eisbuk/shared";
-import { DateTime } from "luxon";
 
 type Firestore = admin.firestore.Firestore;
 
@@ -24,34 +22,13 @@ const relevantCollections = [
   OrgSubCollection.Slots,
 ] as const;
 
-type RelevantCollection = (typeof relevantCollections)[number];
-
-interface EntryExistsPayload {
-  collection: RelevantCollection;
-  exists: boolean;
-}
-
-interface UnpairedCheckPayload {
-  id: string;
-  entries: EntryExistsPayload[];
-}
-
-interface EntryDatePayload extends EntryExistsPayload {
-  date: string | undefined;
-}
-
-interface DateCheckPayload {
-  id: string;
-  entries: EntryDatePayload[];
-}
-
 /**
  * A util used by slot related check cloud function used to find mismatch between slot and attendance entries.
  */
 export const findSlotAttendanceMismatches = async (
   db: Firestore,
   organization: string
-): Promise<SlotSanityCheckReport> => {
+): Promise<SlotAttendanceSanityCheckReport> => {
   const timestamp = DateTime.now().toISO();
 
   // We're only checking for slots/attendances in the last 3 months
@@ -108,52 +85,10 @@ export const findSlotAttendanceMismatches = async (
   return { id: timestamp, unpairedEntries, dateMismatches };
 };
 
-const collectUnpairedDocs = (
-  rec: Record<string, UnpairedDoc>,
-  { id, entries }: UnpairedCheckPayload
-): Record<string, UnpairedDoc> => {
-  // If all entries exist, skip the doc
-  if (entries.every(({ exists }) => exists)) return rec;
-
-  const unpairedDoc = wrapIter(entries)
-    .map(({ collection, exists }) => ({
-      collection,
-      list: exists ? "existing" : "missing",
-    }))
-    ._reduce(
-      (acc, { collection, list }) => ({
-        ...acc,
-        [list]: [...acc[list], collection],
-      }),
-      { existing: [], missing: [] }
-    );
-
-  return { ...rec, [id]: unpairedDoc };
-};
-
-const dateMismatchReducer = (
-  rec: Record<string, DateMismatchDoc>,
-  { id, entries: e }: DateCheckPayload
-): Record<string, DateMismatchDoc> => {
-  const entries = e.filter(({ exists }) => exists);
-  // Keep the first date as reference point
-  const dateRef = entries[0].date;
-  // If dates for all existing entries are the same, skip the doc
-  return entries.every(({ date }) => date === dateRef)
-    ? rec
-    : // If there's a date mismatch, create a [collection]: date record for the doc
-      {
-        ...rec,
-        [id]: Object.fromEntries(
-          entries.map(({ collection, date }) => [collection, date] as const)
-        ) as DateMismatchDoc,
-      };
-};
-
 export const attendanceSlotMismatchAutofix = async (
   db: Firestore,
   organization: string,
-  mismatches: SlotSanityCheckReport
+  mismatches: SlotAttendanceSanityCheckReport
 ): Promise<AttendanceAutofixReport> => {
   const batch = db.batch();
 
@@ -239,65 +174,65 @@ export const attendanceSlotMismatchAutofix = async (
   };
 };
 
-const getSanityChecksRef = (
-  db: Firestore,
-  organization: string,
-  kind: SanityCheckKind
-) => db.collection(Collection.SanityChecks).doc(organization).collection(kind);
-
-const getLatestSanityCheck = async <K extends SanityCheckKind>(
-  db: Firestore,
-  organization: string,
-  kind: K
-): Promise<SanityCheckReport<K> | undefined> =>
-  getSanityChecksRef(db, organization, kind)
-    .orderBy("id", "asc")
-    .limitToLast(1)
-    .get()
-    .then((snap) =>
-      !snap.docs.length
-        ? undefined
-        : (snap.docs[0].data() as SanityCheckReport<K>)
-    );
-
-const writeSanityCheckReport = async <K extends SanityCheckKind>(
-  db: Firestore,
-  organization: string,
-  kind: K,
-  report: SanityCheckReport<K>
-): Promise<SanityCheckReport<K>> => {
-  await getSanityChecksRef(db, organization, kind).doc(report.id).set(report);
-  return report;
-};
-
-interface SanityCheckerInterface<K extends SanityCheckKind> {
-  check(): Promise<SanityCheckReport<K>>;
-  writeReport: (report: SanityCheckReport<K>) => Promise<SanityCheckReport<K>>;
-  checkAndWrite: () => Promise<SanityCheckReport<K>>;
-  getLatestReport: () => Promise<SanityCheckReport<K> | undefined>;
+// #region utils
+interface EntryExistsPayload {
+  collection: OrgSubCollection;
+  exists: boolean;
 }
 
-export const newSanityChecker = <K extends SanityCheckKind>(
-  db: Firestore,
-  organization: string,
-  kind: K
-): SanityCheckerInterface<K> => {
-  const getLatestReport = () => getLatestSanityCheck<K>(db, organization, kind);
-  const writeReport = (report: SanityCheckReport<K>) =>
-    writeSanityCheckReport(db, organization, kind, report);
+interface UnpairedCheckPayload {
+  id: string;
+  entries: EntryExistsPayload[];
+}
 
-  const check = (): Promise<SanityCheckReport<K>> =>
-    ({
-      [SanityCheckKind.SlotAttendance]: findSlotAttendanceMismatches(
-        db,
-        organization
-      ),
-    }[kind]);
+interface EntryDatePayload extends EntryExistsPayload {
+  date: string | undefined;
+}
 
-  const checkAndWrite = () => check().then(writeReport);
+interface DateCheckPayload {
+  id: string;
+  entries: EntryDatePayload[];
+}
 
-  return { getLatestReport, writeReport, check, checkAndWrite };
+const collectUnpairedDocs = (
+  rec: Record<string, UnpairedDoc>,
+  { id, entries }: UnpairedCheckPayload
+): Record<string, UnpairedDoc> => {
+  // If all entries exist, skip the doc
+  if (entries.every(({ exists }) => exists)) return rec;
+
+  const unpairedDoc = wrapIter(entries)
+    .map(({ collection, exists }) => ({
+      collection,
+      list: exists ? "existing" : "missing",
+    }))
+    ._reduce(
+      (acc, { collection, list }) => ({
+        ...acc,
+        [list]: [...acc[list], collection],
+      }),
+      { existing: [], missing: [] }
+    );
+
+  return { ...rec, [id]: unpairedDoc };
 };
 
-type SanityCheckReport<K extends SanityCheckKind> =
-  FirestoreSchema["sanityChecks"][string][K];
+const dateMismatchReducer = (
+  rec: Record<string, DateMismatchDoc>,
+  { id, entries: e }: DateCheckPayload
+): Record<string, DateMismatchDoc> => {
+  const entries = e.filter(({ exists }) => exists);
+  // Keep the first date as reference point
+  const dateRef = entries[0].date;
+  // If dates for all existing entries are the same, skip the doc
+  return entries.every(({ date }) => date === dateRef)
+    ? rec
+    : // If there's a date mismatch, create a [collection]: date record for the doc
+      {
+        ...rec,
+        [id]: Object.fromEntries(
+          entries.map(({ collection, date }) => [collection, date] as const)
+        ) as DateMismatchDoc,
+      };
+};
+// #endregion utils
