+++ packages/functions/src/checks/types.ts
@@ -1,0 +1,3 @@
+import admin from "firebase-admin";
+
+export type Firestore = admin.firestore.Firestore;
